Extend the gbp.conf example to be more complete at default level
Remove the `###` so that the gbp.conf file in its entirety is added to
new packages created with `debmake`. Just adding the file with a comment
about man page didn't help much, and it makes more sense to have the
full example even at basic generation levels.

In the file itself, use defaults that most packages will benefit from.

This is version is also aligned with latest dh_make template for same file.